using Item instead of Rackham using file_set in model rather then generic file
