--- app/actors/curation_concerns/rackham_actor.rb
@@ -1,7 +1,0 @@
-# Generated via
-#  `rails generate curation_concerns:work Rackham`
-module CurationConcerns
-  class RackhamActor < CurationConcerns::BaseActor
-    include ::CurationConcerns::WorkActorBehavior
-  end
-end
