--- app/controllers/curation_concerns/rackhams_controller.rb
@@ -1,7 +1,0 @@
-# Generated via
-#  `rails generate curation_concerns:work Rackham`
-
-class CurationConcerns::RackhamsController < ApplicationController
-  include CurationConcerns::CurationConcernController
-  set_curation_concern_type Rackham
-end
