+++ app/models/file_set.rb
@@ -1,0 +1,4 @@
+# Generated by curation_concerns:models:install
+class FileSet < ActiveFedora::Base
+  include ::CurationConcerns::FileSetBehavior
+end
