--- app/models/generic_file.rb
@@ -1,4 +1,0 @@
-# Generated by curation_concerns:models:install
-class GenericFile < ActiveFedora::Base
-  include ::CurationConcerns::GenericFileBehavior
-end
