--- app/views/curation_concerns/rackhams/_rackham.html.erb
@@ -1,2 +1,0 @@
-<%# This is a search result view %>
-<%= render 'catalog/document', document: rackham, document_counter: rackham_counter  %>
