+++ config/initializers/curation_concerns_config.rb
@@ -1,4 +1,4 @@
 CurationConcerns.configure do |config|
-  # Injected via `rails g curation_concerns:work Rackham`
-  config.register_curation_concern :rackham
+  # Injected via `rails g curation_concerns:work Item`
+  config.register_curation_concern :item
 end
